Fix an issue with blockquotes in wkhtmltopdf
There's a bug in wkhtmltopdf where the use of the 'quotes' CSS rule on the <blockquote> element causes a segfault on Ubuntu.

We don't use this but it is included in the reset-css package. Therefore, this creates a copy of the reset-css stylesheet for PDF use only, and removes the offending `quotes` definition.