Fix broken test for looking up teams by sprints
This had broken because the start dates were not explicitly set for the team, therefore it was falling back to the FactoryBot default of "BASE_START_ON", which in turn was 4 weeks ago.

Because we froze time at the sprint_end_date, the BASE_START_ON was always set to 2022-05-01, which would have broken after that date passed this year. :)