Redirect to the current quarter when it exists
This changes the redirect behaviour of the quarters root page, so that when there is a currently active quarter, we redirect to it.

The existing behaviour of redirecting to the first quarter is maintained as a fallback, when there is no current quarter.

As before, if no quarters exist, we render the view containing a holding message.